requirements.txt
aws-cdk.core
aws-cdk.assertionss
aws-cdk.aws_lambda
aws-cdk.aws_apigateway
cdk-spa-deploy